Chore: Rename test suites to include `.test` extension prefix to prevent @typescript-eslint/unbound-method errors
