+++ src/__tests__/spies.vitest.test.ts
@@ -7,7 +7,7 @@ import {
 	beforeAll,
 	afterAll,
 } from "vitest";
-import {testSuite} from "./test-suites/spies";
+import {testSuite} from "./test-suites/spies.test";
 
 testSuite({
 	describe,
