+++ src/__tests__/window-history.jest.test.ts
@@ -7,7 +7,7 @@ import {
 	beforeAll,
 	afterAll,
 } from "@jest/globals";
-import {testSuite} from "./test-suites/window-history";
+import {testSuite} from "./test-suites/window-history.test";
 
 testSuite({
 	describe,
