+++ src/hooks/__tests__/replace-history-node.jest.test.ts
@@ -9,7 +9,7 @@ import {
 	afterAll,
 } from "@jest/globals";
 import {replaceHistory} from "../replace-history.jest";
-import {testSuite} from "./test-suites/replace-history-node";
+import {testSuite} from "./test-suites/replace-history-node.test";
 
 testSuite({
 	describe,
