+++ src/hooks/__tests__/replace-history-node.vitest.test.ts
@@ -9,7 +9,7 @@ import {
 	afterAll,
 } from "vitest";
 import {replaceHistory} from "../replace-history.vitest";
-import {testSuite} from "./test-suites/replace-history-node";
+import {testSuite} from "./test-suites/replace-history-node.test";
 
 testSuite({
 	describe,
