+++ src/hooks/__tests__/replace-location-missing-global-proxy.vitest.test.ts
@@ -9,7 +9,7 @@ import {
 	afterAll,
 } from "vitest";
 import {replaceLocation} from "../replace-location.vitest";
-import {testSuite} from "./test-suites/replace-location-missing-global-proxy";
+import {testSuite} from "./test-suites/replace-location-missing-global-proxy.test";
 
 testSuite({
 	describe,
