+++ src/hooks/__tests__/test-suites/replace-location-missing-global-proxy.test.ts
